Add RenamingMonitorProcessor
The RenamingMonitorProcessor is a delegating MonitorProcessor which will
change the name attribute of a Monitor by prefixing the values of other
named attributes. This is useful if we want to prefix a Monitor's name
with the value of a few global attributes (like hostname or instance
number) before it is processed by the delegate MonitorProcessor.